app/requirements.txt
Django>=5.0,<6.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
djangorestframework